Develop

+++ social-service/src/main/java/vn/edu/hcmut/social/dto/request/internal/InternalTutorRatingRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class InternalTutorRatingRequest {
     Double averageRating;
-    Integer ratingCount;
+    Long ratingCount;
 }
